refactor: split equipment unit-type validation constraints
